+++ src/Exceptions/InvalidLoaderException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Crwlr\CrawlerExtBrowser\Exceptions;
-
-use Exception;
-
-class InvalidLoaderException extends Exception {}